Set suffix on latest tag for nonroot containers
By default, the docker metadata action treats the `latest` tag differently than
all other tags and does not apply the specified suffix. This resulted in a race
condition where both the root and nonroot builds tried to push the `latest` tag
and the last one won. This sometimes resulted in nonroot build being pushed as
`latest` instead of `latest-nonroot` as desired. Fix is to explicitly specify
that the suffix should apply to the latest tag too.