feat: add support for implicit return
Partially helps with #8408 (but doesn't close it).

In languages such as Julia, Ruby, and Rust, an expression that is
last executed in a function is implicitly that function's return value.

This PR supports matching these implicit returns.

For example, the pattern `return 0` can match functions such as
  ```julia
  f() = 0
  ```
and
  ```julia
  function f()
    0
  end
  ```
and more.

Tested with `make core-test` and added new tests for Julia, Ruby, and Rust.