fix(ci): semgrep ci --pro defaults to legacy
`semgrep` runs `osemgrep`, which for most things switches back to
the python version. However, when `semgrep --pro` is passed, we try
to run `osemgrep-pro`. If this isn't present, we fail.

This is fine for `semgrep scan` users but `semgrep ci` users cannot
respond to the failure as easily. This is why legacy `semgrep ci` will
install semgrep-pro for the user. To resolve this for the osemgrep
entry point, `semgrep ci --pro` will default to legacy semgrep if
`osemgrep-pro` is not already there.

Test plan: remove semgrep-pro such that `semgrep ci --pro` fails.
Then install this version of semgrep and run it again. You should see:

```
➜  single_namespace git:(emma/make-same-namespace-tests-pass) ✗ semgrep ci --pro
Failed to find semgrep-core-proprietary in PATH or in the semgrep package.
You may need to run `semgrep install-semgrep-pro`
Since `semgrep ci` was run, defaulting to legacy semgrep

┌────────────────┐
│ Debugging Info │
└────────────────┘
...
```

The scan should succeed