Generic-to-IL: Fix translation of C++ "foreach" loops
C++ loop `for (int *p : container) { ... }` is encoded as a Generic
'ForIn' node, using a 'ForInitVar' without an initial value. The IL
translation of such Generic AST is flawed, and in fact that seems like
it should be a 'ForEach' instead. So for now we just intercept this
case and translate it as a for-each loop.

Helps PA-3090

test plan:
make test # new test