filter_irrelenvant_rules: Do not filter with literals/regexes
Due to constant folding, a rule may be able to match a string that does
not actually exist as-is in the target file, but that it is produced by
propagating constants. E.g., a rule may succeed at matching "foobar" if
the target code contains `"foo" + "bar"`, but a "foobar" filter would
skip the file!

TODO: clean up

test plan:
TODO