feat: modify origin in `semgrep.dev` metadata for non-semgrep sources (#8719)
When a rule is not downloaded from semgrep.dev by the CLI, strips out
the `semgrep.dev` metadata field if present and the rule also has a
postprocessor. This is to ensure we are able to perform a check against
the origin of the rule in pro so that in normal execution of `semgrep
ci` untrusted postprocessors are not run.

This PR also adds a flag to opt-in to allowing untrusted postprocessors.

Test plan: e2e tests in pro.