ci: configure permissions for the pr-checklist workflow
The PR checklist workflow seems to be failing when dependabot creates a PR

According to

https://docs.github.com/en/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions#changing-github_token-permissions

dependabot defaults to read-only permissions, so I'm configuring the
workflow config to also add write permissions.

I'm not sure how to test this except wait until next week when
dependabot is triggered again, but I think this is low risk and it
seems to be what others have tried and succeeded.