feat(c/c++): update for `tree-sitter-c` 0.20.2 (#8772)
## What:
This PR updates `tree-sitter-c` to 0.20.2 (and `tree-sitter-cpp`
correspondingly), as part of a series of changes to bring C and C++ up
to speed.

## Why:
We want to have better C/C++ support in Semgrep.

## How:
Updated the translation.

Three of the C++ parsing tests fail to parse after this change, because
they don't seem to be valid programs according to the tree-sitter-cpp
grammar. I just changed them slightly to fit into the grammar. If that's
wrong, it's an upstream issue.

## Test plan:
`make test`

Old stats:
```
lang/cpp/stats.json:3
   "global": {
     "name": "*",
     "parsing_rate": 0.9107113211825673,
     "line_count": 19227320,
     "error_line_count": 1716782,
     "file_count": 62388,
     "error_file_count": 26755,
     "total_node_count": 214907024,
     "untranslated_node_count": 375407
   },
```

New stats:
```
lang/cpp/stats.json:3
   "global": {
     "name": "*",
     "parsing_rate": 0.9170075017318946,
     "line_count": 18549050,
     "error_line_count": 1539432,
     "file_count": 60568,
     "error_file_count": 25538,
     "total_node_count": 209934224,
     "untranslated_node_count": 369812
   },
```