Sort findings identically in pysemgrep and osemgrep, text and JSON (#8800)
The order used by osemgrep is: file path, start line, start column, end
line, end column, rule name. This is the order used by pysemgrep. It is
determined by the sorting key defined by `get_ordering_key` in the
`RuleMatch` class (this definition hasn't changed but it wasn't used
everywhere due to some incorrect resorting taking place).

The main benefit of this is that we can now check a bunch of osemgrep
tests by comparing their output to the pysemgrep output without errors
due to the findings being simply in a different order.

I added two osempass tests to check the ordering. A bunch of other test
snapshots changed because I fixed the ordering of pysemgrep to sort the
results by rule ID. They were previously grouped by rule ID but not
sorted by rule ID in JSON. In text output, there was some sorting to
correct for this but it was slightly different than pysemgrep/JSON and
osemgrep so I removed it.

I hope that we can now mark a bunch of tests as osempass. I didn't check
that.

---------

Co-authored-by: Martin Jambon <martin@semgrep.com>