chore: fix multistep rules tests (#8787)
In the semgrep-proprietary repo, one of the multistep tests (secrets)
fails depending on the folder name
of another test (secrets_python_only). This happens because the
`-filter_irrelevant_rules` prefilter uses
a memoized table and the table hmemo is created toplevel in
Analyze_rule.ml, which means it persists
between runs. Currently, this is only a problem for the tests which call
`scan_func_for_osemgrep`, but
we could encounter it in any repeated run of the prefilter on sets of
rules.

Instead, this PR solves this by having the table be passed to the
prefilter instead. When you prefilter a
set of rules, you create your own table (or can choose to run it
unmemoized).

Alternatives considered:
- Refactor `regexp_prefilter_of_rule` to work on sets of rules -->
   this was too much effort given where it's used
- Simply bandaid the problem by clearing the tests in between runs -->
this sets up someone else later to spend another day on the same issue

Test plan: see corresponding semgrep-proprietary PR
https://github.com/returntocorp/semgrep-proprietary/pull/995