test: Java boxing equivalences
I started working on this because I thought that sorting this out was a
prerequisite to fixing something that I needed fixed for the Pro Engine.
However, it turned out not to be. I still think that these equivalences
would be nice to have, but I probably won't work on them in the near
future. However, since I wrote the tests I figured I might as well
commit them.

Test plan: Automated tests