fix: Try new type inference even when target is a name
We have a special case for metavariable matching when the target is a
name with a resolved type. I tried to remove this in #7929 and sadly had
to revert it in #7999.

Ideally we would extend the new, more powerful type inference to be able
to handle all of the cases that are handled here. In the meantime, I'd
like to have matching try both the existing special case and the new,
more general type inference.

I have some upcoming work to untangle boxed and primitive types, as well
as Strings, in Java, to support better matching both in OSS and in the
Pro Engine. This will entail changing how typed metavariables are
matched, and I do not want to have to make changes in two places.

This also changes the behavior of one test, which I have previously
discussed. It probably shouldn't match, but we currently conflate `X`
with `Class<X>` so it does.

Test plan: Automated tests