chore: Show conflicts when running `make dev-setup`
We recently encountered an issue where a dependency in `./dev` needed an
old version of `ppxlib`. However, the current version of `ppx_sexp_conv`
needs a newer version of `ppxlib`. Thus, running `make dev-setup`
silently downgraded the version of `ppx_sexp_conv` that was already
installed. The older version of `ppx_sexp_conv` has a bug that leads to
build failures.

With this change, if we ever add a dependency that conflicts, we'll find
out about it when installing dev dependencies rather than having weird
things happen later.

Test plan: I ran `make setup` before we addressed the conflict and it
failed, with opam reporting the conflict.