fix: permit both CI and scan config IDs for pps
Currently postprocessors are required to pass a validation step in
pysemgrep regarding their origin. This is handled incorrectly in some
cases when we report a config_id in the semgrep ci flow. This resolves
the issue by permitting the config id used for this purpose in addition
to the one associated with registry pro rules.