chore: Get parsing stats running again
For some reason, we are getting segfaults when running parsing stats on
these projects. It doesn't seem to happen when running Semgrep on these
projects, so I am not sure what is going on. We should investigate, but
in the mean time, we would like the parsing stats to start running again
and this was causing the job to fail.

Test plan:

`./run-lang dart` from the `stats/parsing-stats` directory.