fix(ls): some weird git issue (#8869)
When the language server would scan a file before it was saved, it'd
scan a temporary file. This means none of the git filtering stuff would
apply, and so a ton of findings that weren't relevant would pop up. Also
untracked files in a git repo would not be scanned. This fixes both.

## Test Plan
Manually tested it by opening a file that has issues that have been
committed already, then modifying without saving