commons: Add a 'protect' that plays well with timeouts
If a timeout happens in the middle of `finally()`, Semgrep will crash.
This does not happen much (ever?) now because we have very few timeouts,
but we plan to have finer-grained timeouts durint inter-file tainting,
and that makes it more likely to happen.

test plan:
make test
This is actually hard to reproduce, you need to be kind of unlucky,
but it can happen and I have experienced it.