Simplify input to semgrep-core (#8775)
The list of targets discovered by pysemgrep is no longer accompanied by
the list of rules applying to each target. Determining which rule
applies to a target is now done by in OCaml by semgrep-core. This
includes applying the per-rule target exclusions (`include`, `exclude`)
which was already in place and checking that the target's intended
analyzer ("language") is one of those supported by the rule (I had to
add this).

This is meant to reduce the memory needed to load the targets, since
each target path used to come with a list of rule IDs and it could be
large.

Notes:
* I didn't check for the performance of the new implementation.
* There may be more filtering involved than necessary since we're not
clustering rules by language/analyzer like we could and instead we go
through the whole list of rules for each target so select the subset of
applicable rules.
* I didn't get to remove much Python code. It will shrink when we remove
the splitting of tasks by language.

---------

Co-authored-by: Martin Jambon <martin@semgrep.com>