feat(ls): port login to osemgrep (#8899)
## What:
This PR ports `semgrep login` functionality to `osemgrep`, and retires
`pysemgrep`'s LSP code as the entry point for `semgrep lsp`.

## Why:
We would like to officially make `osemgrep` the entry point for LSP!

## How:
Ported code, changed the entry point. There's a lot of `Lwt` wrangling
to be done, because we have to use asynchronous helpers.

## Test plan:
Manually tested that logging in, logging out, and then logging back in
functioned as expected.