test: Fail on empty set of parsing tests (#8904)
https://github.com/returntocorp/semgrep/pull/8773 broke `Common2.glob`
which this relies on. This change makes it so that if we find no parsing
tests for a given language, we error out. This failsafe will make it
more difficult to accidentally stop running the parsing tests in the
future (via changes to `Common2.glob`, or moving the files and making a
typo, or something else).

Test will be red on this until `Common2.glob` is fixed.

Test plan: `make core-test` -> observe a failure (as of October 3)