Catch "missing plugin" errors more widely (#8945)
This moves the spot where we catch "missing plugin" errors occurring
when running Apex rules without the Apex plugin. This was a problem for
rules targeting Apex files for which there was no Apex pattern to parse
(e.g. only `pattern-regex`) which was resulting in a delayed exception
that wasn't caught or no error at all.

Internal ticket:
https://linear.app/semgrep/issue/PA-3173/revisit-pro-rule-skipping-since-its-actually-not-working

---------

Co-authored-by: Martin Jambon <martin@semgrep.com>