feat: add support for implicit return
One more attempt, let's goooo!

Partially helps with #8408 (but doesn't close it).

In languages such as Julia, Ruby, and Rust, an expression that is last
executed in a function is implicitly that function's return value.

This PR supports matching these implicit returns.

For example, the pattern `return 0` can match functions such as
  ```julia
  f() = 0
  ```
and
  ```julia
  function f()
    0
  end
  ```
and more.

This feature can be disabled through the `implicit_return` rule option.

Most of the changes are similar to
https://github.com/returntocorp/semgrep/pull/8826 with some notable
changes:

* Added `CFG_build.cfg_of_fdef` which is wraps the logic of creating a
  CFG, mark nodes, and re-creating the CFG for languages that support
  implicit return. There are 4 call sites for this function now. I
  thought it made sense to include all the logic within the function
  and not the call sites.

* Updated `AST_to_IL.call_generic`. It was calling `AST_generic.e`
  every time to create a new expression node, so any changes to
  is_implicit_return is not persisted across CFG builds for call nodes
  that were created by `call_generic`.

* Removed an unneeded parameter from `AST_to_IL.cond_with_pre_stmts`.

* One taint_if maturity test in Elixir is failing after this
  change. However, the test was previously passing for the wrong
  reasons. I looked at the CFG in HEAD, and it was already wrong, but
  the tests just happened to pass. So I've marked the missing match as
  a TODO for now, and it will fixed later, separately.

Tested with `make core-test` and added new tests for Julia, Ruby, and Rust.

As for Elixir, the language support is still experimental. I'll add
more tests later as it matures.