jsonnet workflow improvements (#8991)
This PR does a few things:
- Updates the workflows Makefile to additionally depend on the libsonnet
files, so that a change to a lib regenerates the yamls.
- Updates `ojsonnet` to only output one trailing newline (the double
newlines was triggering the "fix newlines" pre-commit check which made
committing workflow changes a pain). I did this by adding an optional
`newline` arg to `pr` and `pr2`. LMK if folks think there's a better way
to do this.
- Exempts the generated yamls from the prettier pre-commit check. IMO
the ojsonnet output is readable enough and removing the friction of
having to prettier all your changes is nice
- Regenerates the yamls in response to the changes above

test plan:
- tests pass
- structural diffs are same