filter_irrelevant_rules: Do not filter with every literal/regex (#8767)
Due to constant folding, a rule may be able to match a string that does
not actually exist as-is in the target file, but that it is produced by
propagating constants. E.g., a rule may succeed at matching "foobar" if
the target code contains `"foo" + "bar"`, but a "foobar" filter would
skip the file!

Closes PA-3110

test plan:
make test # added new tests
Plus ran p/default-v2 on 46 repos and on average runtime and memory
usage were essentially the same. For five repos, there was a more
significant slowdown of about 20%, due to more files being analyzed.
There is a large increase in "partially scanned files" because more
files are being analyzed, and we encounter some parsing errors on them.
This is expected and not bad per se.