feat!(secrets): Implement masking in pysemgrep
Currently masks everything past the first 20% of the finding in text
output, as well as metavariables by replacing the content with *s. This
is to avoid showing secrets in logs.

This also currently has some changes to the syntatic ID hash. To wit, we
would update it to (1) hash the existing hash with BLAKE2; (2) suffix
the updated version with /v2; (3) on ingesting any non-suffixed hash,
rehash with BLAKE2. The intention of this is to allow us to update the
existing hashes while moving to a more secure hashing algorithm (See
also SEC-396 in Linear). I am not sure of all possible consequences here
and may need to revise slightly depending on what coupling this would
have with app. Notably, we may currently run into issues where versions
of the CLI older than this commit get the 'v2' syntatic ID and then
cannot perform deduplication correctly in CI. I don't know how much of
an issue, if at all, that would be.

It may be possible for us to use a different "syntatic ID" for secrets,
but this doesn't address some issues with the current usage of Murmur3.