chore: cleanup and add tests for prefilter
Cleans up after #9010

Improve the type so that we only create a cache when it will actually
be used by the filter-irrelevant-rules optimization. Also adds benchmark
tests to catch a similar error in the future.

Test plan: make test

Also, run semgrep OSS and print to confirm that rules are being filtered