fix: Refuse to run out of date semgrep-core-proprietary (#9033)
Currently, if you install semgrep-core-proprietary, then upgrade
Semgrep, then try to run the Pro Engine, Semgrep will use the old
version of semgrep-core-proprietary. This often results in crashes.
Instead, we should ask users to update semgrep-core-proprietary.

Since the semgrep-core-proprietary binary does not know what version of
Semgrep it has been built for, the simplest way to accomplish this is to
add a file alongside the binary indicating which version of Semgrep
installed it. When we execute semgrep-core-proprietary, we can look at
that file. If it's different than the current version of Semgrep, we
refuse to run the old binary.

Fixes #8873

Test plan:

Rebase onto the v1.44.0 tag to ensure compatibility with the latest
released version of `semgrep-core-proprietary` and run the following
manual tests:

- `semgrep scan --pro --config rules.yaml`
  - With no semgrep-core-proprietary installed (nice error message)
  - With semgrep-core-proprietary installed (success)
- With `pro-installed-by.txt` changed to a different version (nice error
message)
  - With `pro-installed-by.txt` removed (nice error message)
- `semgrep scan --pro --config rules.yaml --legacy`
- With no semgrep-core-proprietary installed (bad error message, but
same as before)
  - With semgrep-core-proprietary installed (success)
- With `pro-installed-by.txt` changed to a different version (bad error
message, same as when it's not installed)
  - With `pro-installed-by.txt` removed (same bad error message)
- `semgrep scan --config rules.yaml`
  - With no semgrep-core-proprietary installed (works fine)
  - With semgrep-core-proprietary installed (works fine)
- With `pro-installed-by.txt` changed to a different version (works
fine)
  - With `pro-installed-by.txt` removed (works fine)
- `semgrep ci` with Pro toggle on
  - With no semgrep-core-proprietary installed (downloads and runs)
  - With semgrep-core-proprietary installed (runs)
- With `pro-installed-by.txt` changed to a different version (downloads
and runs)
  - With `pro-installed-by.txt` removed (downloads and runs)
- `semgrep ci` with Pro toggle off
  - With no semgrep-core-proprietary installed (runs)
  - With semgrep-core-proprietary installed (runs)
  - With `pro-installed-by.txt` changed to a different version (runs)
  - With `pro-installed-by.txt` removed (runs)