chore: Delete test-semgrep-pro-latest CI job
This job was created as a stopgap so we could avoid shipping breakages,
back when we served only a single version of the
`semgrep-core-proprietary` binary no matter what version of Semgrep was
requesting it.

Fortunately, now we serve the appropriate `semgrep-core-proprietary`
binary for the requesting version of Semgrep, so we no longer need to
care about compatibility between versions.

Furthermore, this job wasn't actually testing what we wanted it to test
because it used the current `returntocorp/semgrep:develop` image, but
didn't wait for that image to get updated based on the contents of the
current commit.

Test plan: CI