Unrevert "fix: Refuse to run out of date semgrep-core-proprietary" (#9070)
This reverts commit efc8d1594f603ec21ae0da710fc7fc3578e2085d.

I reverted this because I was not able to address all of the issues
before the release. Fortunately, now I have.

In addition to the tests laid out in the original PR (#9033), I've now
tested the nonroot docker image by building it locally and running it
while passing `--user semgrep`, and I have tested Homebrew by following
these instructions
https://docs.brew.sh/How-To-Open-a-Homebrew-Pull-Request, pointing my
local Homebrew formula for Semgrep to the `nmote-brew-test-1.46.0` tag,
installing Semgrep Pro, and verifying that it is accompanied by the
version stamp and that it runs successfully.

I have also migrated a couple problematic CI jobs in
`semgrep-proprietary` over to `semgrep install-semgrep-pro
--custom-binary` instead of directly moving the just-built binary to the
place where Semgrep expects it.