devops: use slackapi/slack-github-action@ instead of curl
This was inspired by a recent PR from jonas in semgrep-app.

The advantage of curl is that we can specify here the destination.
With curl it's a two steps process where we need to configure
somewhere else the channel to send to

test plan:
activate workflow and observe notifications in #team-semgrep-core