commons: Add a 'protect' that does not raise 'Finally_raised' (#8900)
With `Fun.protect ~finally work`, if a timeout happens in the middle of
`finally()`, it will raise a `Finally_raised` exception rather than a
`Timeout` exception, so it will be handled as a fatal error rather than
as a timeout. This does not happen much (ever?) right now because
timeouts are rare, but we want to have finer-grained timeouts during
inter-file tainting, particularly during taint-signature inference. When
there are more timeouts, it becomes more likely than one coincides in
the middle of a `finally`. (I have experience this running on
_elasticsearch_ with fine-grained timeouts.)

For now, we just drop `Finally_raised` exceptions and instead re-raise
whatever exception was raised during `finally`.

Eventually we may want to write a more robust `protect` function that
blocks `SIGALRM`, but it seems that first we may need to review every
single use of `protect`. (We tried blocking `SIGALRM` and ran into some
weird problems, see code comments.)

test plan:
Add fine-grained timeouts to inter-proc tainting and run with --pro on
elasticsearch