tainting: Add `exact` option to sources and santizers
This option makes the identification of sources (sanitizers) stricter,
or more precise, similar to what we already for for sinks. But it appears
that more rules rely on the "non-exact" semantics of sources (sanitizers),
so for now we keep the default as it was.

Related-to: 4020470dd7c ("tainting: Make identification of sinks more precise (#7215)")

Closes #5897
Closes PA-1752

test plan:
make test # new tests