fix(ls): Don't scan every file on startup, and don't use pro engine rules (#9079)
This PR makes it so an entire repo is not scanned when the extension
starts. This decision was made for performance reasons, as this causes
heavy computation, when it's usually not used. My justification is that
editors like intellij don't even provide a way to see problems in
non-open files, and extensions like ruff don't scan until a file is
opened, so it feels unnecessary for us to do so too. The command to scan
the entire workspace still exists though.

This PR also makes it so rules requiring the pro engine, or unported
code (SCA, Interfile, Secrets) are not run. This should reduce noice and
improve performance.

## Test Plan
LS e2e tests were updated accordingly