Fix --severity regression
This closes #9062

The regression was introduced in #8958 when switching
from Enum to ATD-defined MatchSeverity in semgrep_output_v1.atd
Not sure why mypy didn't complain about the
  `rule.severity.value in severity`
where we compare a MatchSeverity and a string.

Note that osemgrep --severity was working!

Anyway, this is fixed now.

test plan:
added test in test_severity.py
'make e2e'.