feat: Secrets ignores semgrepignore (#9125)
This change modifies core and cli to bypass semgrepignore. This isn't
configurable for secrets right now, eventually we would like there to be
a notion of a section in the semgrep ignore. Instead we target all
configuarable files. This effectively implements the MVP for public beta
in [this scope
doc](https://www.notion.so/Semgrep-Secrets-File-Targeting-bee9d620d54b4ddf81c8f9cff203ba33).
 
Testing is done in [this semgrep-pro
PR](https://github.com/returntocorp/semgrep-proprietary/pull/1093).