fix: incorrect metadata change if multiple matches (#9149)
Reported by @LewisArdern. We incorrectly shared a metadata blob on a
per-rule basis, rather than on a per-finding basis. This resulted in
updates to a finding's metadata being incorrectly applied to all
findings for a given rule, such that later findings would modify the
metadata of earlier ones. Now clones the metadata when modifying so that
we can modify independently.

Test plan: requires additional e2e tests in pro.