chore: Fix release process to avoid malformed Python wheels
The 1.47.0 release failed upon uploading to pypi because the wheel
included a DESCRIPTION.rst file which was actually our README. The
README is markdown, which is incompatible with the rst format. Previous
versions did not have this file at all, and instead the README was
correctly included in the METADATA file in the wheel, which allowed it
to be displayed correctly on pypi.

I believe that the workflow changes in #9081 caused this, but I am not
certain.

This PR does two things:

- Runs `twine check` on PRs. This catches the issue with malformed
  packages, as evidenced by the test plan.
- Upgrades to the latest version of setuptools and other Python
  packages. This solves the issue with malformed packages getting
  generated.

Test plan:

I verified that `twine check` catches the malformed wheel at PR time:
https://github.com/semgrep/semgrep/actions/runs/6748222634/job/18346377598

This run was before I fixed the underlying issue, and it is the same
error that caused the release to fail.

Then, I verified that updating the various Python packages fixed the
issue with the malformed wheel, shown by the fact that even with the
`twine check` step, the CI jobs on this PR are green.