fix: handle definition expressions in ast_to_il
Partially helps: [PA-3185](https://linear.app/semgrep/issue/PA-3185).

This allowed tainting to find a match in taint_if.ex.

Previously, definition expressions were not handled, and it resulted
in an incomplete CFG for one of the functions.

This PR results in a complete CFG, so tainting was able to properly
analyze that function.

This