chore: remove node dependencies from engine (#9187)
## What:
This PR removes `node_js_shared` from the engine.

## Why:
It's not necessary, we only need those dependencies for the Unix calls,
which are not called from just the scans themselves in the engine. This
will cause Turbo Mode to emit an error in the Playground.

## Test plan:
Observe that by running the App locally, with a newly-built version of
the JS Engine, will no longer emit a warning for `node:process`.