Raise default timeout from 2s to 5s (#9167)
Ran the full compare-semgrep benchmark and it did not show meaningful
differences. In a sense, that means 5s does not seem to hurt perf and it
also gives Semgrep extra flexibility when it is needed. Based on
user/customer observations, we also believe it should be a better
timeout. Let's try.

Helps FIND-1270

---------

Co-authored-by: Clara McCreery <chmccreery@gmail.com>