Make validation status different for deduplication
Previously we could reasonably have findings which were equal under cli
match id but had different validation states. The common case for this
is a secret comprising a part A and a part B. Such A+B secrets report a
finding at one part (wlog A) and the deduplication logic does not
consider A+B and A+B' to be distinct on account of there not being a
clear way to tell B and B' apart when they are not used in the message.

However, we don't want to consider most findings distinct only on the
basis of non-consumed metavariables since this results in many findings
which are duplicative. We should consider either this method, or
considering all metavariable used in the validators (if any). In either
case, we need to come up with a clear way to disambiguate findings,
since in the case where we have A+B, a putative secret, and A+B', ibid.,
and we report findings at A, it is not clear from the output what
matters in cases where A+B is valid and A+B' is not (e.g.).

It may be worthwhile considering making `anywhere` result in "additional
context" similar to a taint trace, which is shown for all findings. That
said, this may be less than ideal in the main view in the app.
(Consider, for instance, A0, A1, A2, A3 and the same for B0-3, where
A0+B0 is valid and so on---then we would flag 4 valid secrets and 12
"invalid" ones where the invalid ones are irrelevant)