feat: `semgrep-internal-pattern-anywhere` (#9214)
See also <https://www.notion.so/semgrep/semgrep-internal-patterns-anywhere-bd56b763e470494bba9213523c45a5cf?pvs=4>

This introduces an additional pattern key,
`semgrep-internal-pattern-anywhere` (`anywhere`, in the "new" syntax),
which results in the subpattern being combined without respect to range
at a `patterns` (bzw. `all`) or similar. I.e., a pattern like

```yaml
patterns:
- pattern: A
- semgrep-internal-pattern-anywhere: { pattern: B }
```

will result in a finding at occurences of `A` if `B` is present anywhere
(in the same file). Since the combination is without respect to range,
the range associated with the pattern under
`semgrep-internal-pattern-anywhere` is discarded and matches only occur
(range-wise) at A. Importantly, metavariables are still treated
"normally" in that they are required to unify, so

```yaml
patterns:
- pattern: $F(...)
- semgrep-internal-pattern-anywhere:
  - pattern: |
      def $F(...):
          ...
```

would match all calls to all functions that were defined in the same
file.

Test plan:
- Additional tests added

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>