devops: Add semgrep-comparison trigger to tests (#9251)
Depends on https://github.com/semgrep/semgrep/pull/9243.

This PR adds the trigger for the `semgrep-comparison` Argo Workflow to
`tests.yml`. This will trigger a new check-run on PRs which runs the
`semgrep-comparison` benchmarks against `develop`.

### Test plan

Run the `trigger-semgrep-comparison-argo` workflow with
`workflow_dispatch`. This is an adequate approximation of how it will
behave on a PR; however, there may be some GHA-isms that I missed which
may require follow-up PRs.

I could test `tests.yml` with `act`, but that requires a lot of
additional setup to mimic our secrets, environments, etc.


![image](https://github.com/semgrep/semgrep/assets/2374948/c3e8462f-a567-473b-86f4-bd077e62a91c)


![image](https://github.com/semgrep/semgrep/assets/2374948/11ae4d70-5279-43fb-879c-8a67469da455)


Oh, actually this PR ran the changes to the workflow. See the job
`semgrep-compare-github-cmmn7` below!


![image](https://github.com/semgrep/semgrep/assets/2374948/68f50951-fedd-441f-ae0f-1954e8c1b4d3)