tests: Make mini-benchmarks workflow less noisy (#9246)
We get a lot of "The whole benchmark suite is too slow" comments in PRs
and we tend to ignore them, and this is just "training" us to ignore
them too when the perf deviation is larger and actually worth
investigating.

So, let's raise the global threshold from 7% (~1 seconds slower) to 15%
(~2 seconds slower). Given that we run these benchmarks in GHA, that the
entire benchmark is ~1 second slower can easily be attributed to noise.

Also, this is less critical nowadays since we have a more comprehensive
benchmark in the form of an Argo workflow that we run nightly and
on-demand. (And hopefully it will soon run on CI too.)

test plan:
No "The whole benchmark suite is too slow" comment in this PR