feat: cache_disk for lwt
Uses the same cache as normal but allows a `t Lwt.t` to be returned from
the function and then cached when `t` is actually yielded. This cannot
be done witht the current cache function since we would attempt to cache
the `t Lwt.t` when returned, but we want to actaully be saving the `t`
(inter alia, because we can't readily serialize `t Lwt.t` with
`Marshal`).

This is helpful for caching network requests we want to leverage lwt
for.