fix(osemgrep): Implement `--scan-unknown-extensions` for osemgrep
The `--scan-unknown-extensions` option asks semgrep to scan files that
are explicitly passed and have unknown extensions with all the given rules.

To handle this in osemgrep, when we perform file targeting, for each file,
if it's explicitly passed and `--scan-unknown-extensions` is true, the file
is marked as an explicitly passed file that we should consider checking for
an unknown extension. Later, when we partition the jobs by language, we add
on the files with unknown extensions to each job.

Test plan: in addition to the existing automated tests, I also ran
```
osemgrep --scan-unknown-extensions --strict --config rules/eqeq-python.yaml --json targets/basic/stupid_no_extension targets/basic/nosem.py targets/nosem_invalid_id/ --experimental
```

and modified `eqeq-python.yaml` to have a javascript rule with the pattern `$F(...)`.

I expected findings on `targets/basic/stupid_no_extension` and `targets/nosem_invalid_id/nosem_invalid_id.js`.
I expected a partial parse error on `targets/basic/stupid_no_extension` but not on `targets/basic.nosem.py`,
since we attempt to scan the former with both python and javascript but not the latter.