feat(js/parsing): parsed integers (#9197)
## What:
This PR introduces the `Parsed_int` module, and replaces many instances
of `int option wrap` in our various ASTs with `Parsed_int.t`.

## Why:
We use `int`s in many places.

Other than its obvious usage for counting, for indices, and for unique
identifiers, we must have a representation of integers for when we
represent the integers in a target program. This so far has been just a
normal `int` type, as OCaml's normal representation for integers.


### JS
`int`s are no problem when running Semgrep's OCaml code natively. This
is no longer the case, because of our efforts to port Semgrep to
Javascript (JSCaml). This leads us to representation issues, because
OCaml's 31/63-bit ints conflict with those of Javascript's 64-bit
integers, and can lead to weird under/overflow situations. As such, we
are looking for a solution which can make Semgrep code still workable in
JS.

### Types and Abstraction
Our usage of `int`s as representations for integers in a target program
is different in kind from those uses of integers as a number. These are
the same type in our code, but actually have different use cases and
different intentions. It would be nice if we could invent an abstraction
to distinguish these two use cases.

## How:
This PR implements a virtual module `Parsed_int`, which is a type that
distinguishes integers _as a number_ from integers _as a symbol_, which
is the primary use case for parsed integers.

## Test plan:
Automated tests.

Closes PDX-147

---------

Co-authored-by: pad <pad@r2c.dev>