fix: tainting: Remove Pro-hack for `lval = new ...`
The hack consisted in not cleaning the taint in `lval` when the RHS was
a `new`-expression. This is because in Pro we are tainting `lval` by
side-effect and we did not want Semgrep to clean that taint.

In principle, this hack could have been removed since PR #9263, once we
were explicitly checking for side-effectful updates to `lval` and only
cleaning `lval` if there were none. But PR #9263 was only checking
changes to `lval` itself, and we have to check changes to `lval` as well
to any of its extensions `lval.a1. ... .aN`.

Once PR #9263 is fixed, this hack for `new` is no longer needed and, in
fact, it was causing FPs in double-free rules, when an object is both
allocated and freed inside a loop.

Fixes: e93afb54fb7 (fix: tainting: Do not clean LHS if there were side-effects (#9263)")

test plan:
make test # new test