fix: tainting: Remove Pro-hack for `lval = new ...` (#9334)
The hack consisted in not cleaning the taint in `lval` when the RHS was
a `new`-expression. This is because for `lval = new ...` we are tainting
`lval` by side-effect, and we do not want Semgrep to clean that taint
afterwards.

In principle, this hack could have been removed after PR #9263, once we
were explicitly checking for side-effectful updates to `lval` and only
cleaning `lval` if there were none. But PR #9263 was only looking for
changes to `lval` itself, and we have to look for changes to `lval` as
well as to any of its extensions `lval.a1. ... .aN`. If the taint of
e.g. `lval.x` has been updated, we should also not clean `lval`.

Once PR #9263 is fixed, this hack for `new` is no longer needed and, in
fact, it was causing FPs in double-free rules, when an object is both
allocated and freed inside a loop.

Fixes: e93afb54fb7 (fix: tainting: Do not clean LHS if there were side-effects (#9263)")

Closes PA-3283

test plan:
make test # new test
And CI workflow 'check-semgrep-pro' must pass.