tainting: Make identification of sinks (even) more precise
Improves on PR #7245 to detect a lot more cases where the intent of a
sink specification is to declare the argument of a function call as the
sink.

This is to avoid that code like `sink(ok1 if tainted else ok2)` will
generate a finding.

Eventually we may add an `exact:` option to sinks as we did for sources,
but that is out of the scope of this PR.

Follows: d09dc5ab050 ("tainting: Make identification of sinks more precise (#7245)")

Closes PA-3284

test plan:
make test # new test