cli: ⏳ Mitigate Registry Latency Pains (#9274)
## Description
We can observe request latency of 4-6 secs in tail cases for fetching
rules from the registry. To offset the pain, we can add a loading
indicator (e.g. a spinner) in our CLI to improve the ux.

**Addresses**: [GROW-79](https://linear.app/semgrep/issue/GROW-79)

### Demo


https://github.com/semgrep/semgrep/assets/5779832/ceccd65f-9635-40c5-bf14-b4b8db372049

#### Implementer Notes
- Our spinner logic in the rule fetching context is coupled with our
single network request and is expect to only wait as long as the pending
request
- Our spinner logic in `login` purposefully intends to act as a busy
loop to prevent hammering our `login` backend endpoint as we wait for
the user to activate their token within a wait + retry loop
- In a future PR we can elect to update and standardize the console
spinner logic used for `semgrep login`

### Test Plan
- [x] validated spinner shows for registry fetch
- [x] validated spinner does not show for local rules (or non-network
rule fetch request)
- [x] CI test pass

---------

Co-authored-by: Andre Kuhlenschmidt <andre@semgrep.com>
Co-authored-by: ajbt200128 <austin@cutedogs.org>
Co-authored-by: Cooper Pierce <cooper@semgrep.com>