Add test subcommand to osemgrep (#9247)
This compiles, but does not work yet (since (a) the semgrep_core is not
hooked yet, and (b) the output is not finished). There's also lacking
the "fixtest_filenames" and "fixtest" in general.

The reason for this PR is that it is safe to merge at any time, and I
appreciate feedback on what is there (and whether parts should be
improved). It took quite some time since the internal data structures
and output structures are not typed in test.py (neither are part of
semgrep_interfaces as far as I could see).